refactor: Refactor components
