+++ src/components/common/footer/styles.ts
@@ -1,0 +1,18 @@
+import styled from "@emotion/styled";
+
+export const FooterContainer = styled.div`
+  display: flex;
+  padding: 40px 16px;
+  background-color: rgb(39, 38, 44);
+  color: white;
+  z-index: 10;
+  justify-content: center;
+`;
+
+export const FooterWrapper = styled.div`
+  width: 100%;
+  max-width: 1280px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+`;
