+++ src/components/common/header/sign-in-up-button/styles.ts
@@ -1,0 +1,9 @@
+/* eslint-disable import/prefer-default-export */
+import styled from "@emotion/styled";
+
+export const Divider = styled.div`
+  width: 1px;
+  height: 24px;
+  background-color: #e0e0e0;
+  margin: 0px 16px 0px 8px;
+`;
