+++ src/components/common/logo-slider/logo-slide/index.tsx
