+++ src/components/common/logo-slider/styles.ts
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled";
+
+export const SliderContainer = styled.div`
+  overflow: hidden;
+`;
+
+export const SliderArea = styled.div`
+  display: flex;
+`;
+
+export const LogoArea = styled.div`
+  display: flex;
+`;
