+++ src/components/common/ordering-button-group/styles.ts
@@ -1,0 +1,10 @@
+/* eslint-disable import/prefer-default-export */
+import styled from "@emotion/styled";
+
+export const OrderingButtonGroupContainer = styled.div`
+  display: flex;
+  justify-content: right;
+  width: 100%;
+  max-width: 1280px;
+  margin-bottom: 16px;
+`;
