+++ src/components/common/pagination/styles.ts
@@ -1,0 +1,8 @@
+/* eslint-disable import/prefer-default-export */
+import styled from "@emotion/styled";
+
+export const PaginationContainer = styled.div`
+  display: flex;
+  justify-content: center;
+  margin: 16px 0;
+`;
