+++ src/components/home/home-volunteer-list/home-volunteer-list-item/styles.ts
@@ -1,0 +1,13 @@
+import styled from "@emotion/styled";
+
+export const CardContainer = styled.div`
+  position: relative;
+`;
+
+export const CardContent = styled.div`
+  position: absolute;
+  bottom: 0px;
+  left: 16px;
+  color: white;
+  z-index: 1;
+`;
