+++ src/components/home/home-volunteer-list/styles.ts
@@ -1,0 +1,24 @@
+import styled from "@emotion/styled";
+
+export const HomeVolunteerListContainer = styled.div`
+  height: 400px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 16px;
+  background-color: var(--wave-color);
+`;
+
+export const ListItemContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  max-width: 1280px;
+`;
+
+export const VolunteerCardContainer = styled.div`
+  width: 100%;
+  display: flex;
+  justify-content: space-around;
+`;
