+++ src/components/home/univ-logo-slider/styles.ts
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+import styled from "@emotion/styled";
+
+export const UnivLogoSlideContainer = styled.div`
+  text-align: center;
+  padding: 60px 0px;
+`;
