+++ src/components/login/login-container/styles.ts
@@ -1,0 +1,28 @@
+import styled from "@emotion/styled";
+
+export const LoginItemContainer = styled.div`
+  margin-top: 60px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+
+export const LoginHeader = styled.div`
+  display: flex;
+  position: relative;
+  width: 100%;
+  max-width: 432px;
+  justify-content: center;
+  margin-bottom: 16px;
+`;
+
+export const RegisterContainer = styled.div`
+  display: flex;
+  margin: auto 0 0 auto;
+  align-items: baseline;
+`;
+
+export const HeaderText = styled.h2`
+  margin-top: 0;
+  margin-bottom: 24px;
+`;
