+++ src/components/post/image-uploader/styles.ts
@@ -1,0 +1,20 @@
+import styled from "@emotion/styled";
+
+export const ImageUploaderContainer = styled.div`
+  display: flex;
+  overflow-y: auto;
+`;
+
+export const ImageUploaderButton = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 150px;
+  min-width: 150px;
+  height: 150px;
+  min-height: 150px;
+  border: 3px dashed lightgray;
+  color: lightgray;
+  border-radius: 10px;
+  cursor: pointer;
+`;
