+++ src/components/post/milkdown-editor/styles.ts
@@ -1,0 +1,12 @@
+import styled from "@emotion/styled";
+
+export const MilkdownContainer = styled.div`
+  display: flex;
+  position: relative;
+  justify-content: center;
+`;
+
+export const MilkdownEditorContainer = styled.div`
+  width: 100%;
+  max-width: 1280px;
+`;
