+++ src/components/post/volunteer-tool-bar/styles.ts
@@ -1,0 +1,10 @@
+import styled from "@emotion/styled";
+
+export const VolunteerToolBarContainer = styled.div`
+  display: flex;
+  justify-content: space-between;
+`;
+
+export const DateTimeContainer = styled.div`
+  display: flex;
+`;
